util/disk_cache: pass timestamps as a blob
We no longer need a string, so can just pass the timestamps as a blob to
reduce amount of data that needs to be hashed into the keys. The
argument is also renamed to version_blob because the cache doesn't
really care if it's timestamps, build id or something else.

Signed-off-by: Grazvydas Ignotas <notasas@gmail.com>